# api/specifications/vitanexus_api_spec.yaml
openapi: 3.0.3
info:
  title: VitaNexus API
  description: |
    REST API for VitaNexus Health Assurance Cooperative.

    Core API for member management, health scoring, financial operations,
    and third-party integrations (wearables, EHR, claims processors).

    **Security**: All endpoints require OAuth 2.0 authentication with JWT tokens.
    PHI data is encrypted in transit (TLS 1.3) and at rest (AES-256).

    **Rate Limiting**: 1000 requests/minute per API key for standard tier.

    **HIPAA Compliance**: All PHI access is logged in audit trail.

  version: 1.0.0
  contact:
    name: VitaNexus API Support
    email: api-support@vitanexus.com
  license:
    name: Proprietary
servers:
  - url: https://api.vitanexus.com/v1
    description: Production server
  - url: https://api-staging.vitanexus.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development
security:
  - OAuth2:
      - read:members
      - write:members
      - read:health_data
      - write:health_data
      - read:financial
      - write:financial
...
paths:
...
  /wearables/connect:
    post:
      tags:
        - Wearables
      summary: Connect wearable device
      description: Initiate OAuth flow to connect member's wearable device
      operationId: connectWearable
      security:
        - OAuth2:
            - write:health_data
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - member_id
                - device_type
              properties:
                member_id:
                  type: string
                  format: uuid
                device_type:
                  type: string
                  enum: [fitbit, apple_health, garmin, whoop, oura]
                callback_url:
                  type: string
                  format: uri
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  authorization_url:
                    type: string
                    format: uri
                  state:
                    type: string
                    description: OAuth state parameter for verification
  /wearables/callback:
    post:
      tags:
        - Wearables
      summary: OAuth callback handler
      description: Handle OAuth callback from wearable provider
      operationId: wearableOAuthCallback
      security:
        - OAuth2:
            - write:health_data
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - state
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        '200':
          description: Connection established
          content:
            application/json:
              schema:
                type: object
                properties:
                  member_id:
                    type: string
                    format: uuid
                  device_type:
                    type: string
                  connected:
                    type: boolean
                  last_sync:
                    type: string
                    format: date-time
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.vitanexus.com/oauth/authorize
          tokenUrl: https://auth.vitanexus.com/oauth/token
          scopes:
            read:members: Read member information
            write:members: Create and update members
            read:health_data: Read health data including PHI
            write:health_data: Write health data
            read:financial: Read financial data
            write:financial: Submit claims and process payments
...